avutil/error: fix AV_ERROR_MAX_STRING_SIZE type

+++ sys/src/avutil/error.rs
@@ -46,7 +46,7 @@ pub const AVERROR_HTTP_NOT_FOUND:    c_int = FFERRTAG!(0xF8, b'4', b'0', b'4');
 pub const AVERROR_HTTP_OTHER_4XX:    c_int = FFERRTAG!(0xF8, b'4', b'X', b'X');
 pub const AVERROR_HTTP_SERVER_ERROR: c_int = FFERRTAG!(0xF8, b'5', b'X', b'X');
 
-pub const AV_ERROR_MAX_STRING_SIZE: c_int = 64;
+pub const AV_ERROR_MAX_STRING_SIZE: size_t = 64;
 
 #[inline(always)]
 pub unsafe fn av_make_error_string(errbuf: *mut c_char, errbuf_size: size_t, errnum: c_int) -> *mut c_char {
